+++ app/lib/screens/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:menstrudel/widgets/basic_progress_circle.dart';
-import 'package:menstrudel/widgets/log_period.dart';
+import 'package:menstrudel/widgets/dialogs/log_period_dialog.dart';
 import 'package:menstrudel/models/period_logs.dart';
 import 'package:menstrudel/models/period.dart';
 import 'package:menstrudel/database/period_database.dart';
